backend/requirements.txt
flask
flask-socketio
eventlet
ultralytics
torch
pillow
numpy
